requirements.txt
anyio==3.6.1
apturl==0.5.2
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
async-generator==1.10
attrs==21.4.0
awscli==1.22.34
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
boto3==1.24.17
botocore==1.27.17
Brlapi==0.8.3
bs4==0.0.1
cachetools==5.2.0
certifi==2020.11.8
cffi==1.15.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==7.1.2
cmake==3.22.4
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==37.0.2
cssselect2==0.6.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.8
django-xhtml2pdf==0.0.4
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
environ==1.0
face-recognition-models==0.3.0
fastapi==0.71.0
fasteners==0.14.1
feedparser==6.0.10
Flask==1.1.2
fonttools==4.33.3
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
grpcio==1.46.1
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
imutils==0.5.3
itsdangerous==1.1.0
jeepney==0.7.1
Jinja2==2.11.2
jmespath==1.0.1
jsonify==0.5
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.5.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.2.1
netifaces==0.11.0
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.64
oscrypto==1.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.5
paramiko==2.9.3
pexpect==4.8.0
pgpdump==1.5
Pillow==9.0.1
pipreqs==0.4.11
protobuf==4.21.2
psycopg2==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.9.0
Pygments==2.11.2
PyGObject==3.42.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyPDF3==1.0.6
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.1.4
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
python-apt==2.3.0+ubuntu2
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
PythonTurtle==0.3.2
pytrends==4.8.0
pytweening==1.0.4
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.0
roman==3.3
rsa==4.8
s3transfer==0.6.0
screen==1.0.1
SecretStorage==3.3.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.0.1
sqlparse==0.3.1
starlette==0.17.1
svglib==1.4.1
systemd-python==234
termcolor==1.1.0
texttable==1.6.4
tinycss2==1.1.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
ttystatus==0.38
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
uritools==4.0.0
urllib3==1.26.1
usb-creator==0.3.7
uvicorn==0.16.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==1.0.1
wsproto==1.1.0
xdg==5
xhtml2pdf==0.2.8
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
